Change in ParameterEstimation API to allow for separation of parameters. Previously, the API only allowed for generation of a full factor. However, many interesting distributions will have far fewer parameters, and the values of those parameters are of interest in themselves.
Additionally, the responsibility for collection of sufficient
statistics has been shifted from ParameterEstimation to
SharedParameters. At this ponit, they are not shared parameters as
much as counts.  In cases where the sufficient statistics
are smaller than the counts, this will require more memory but
will likely execute faster. The worst case scenario is that the
counts take up as much space as the factor graph, so this is not
a huge burden.